+++ openshift-maven-plugin/plugin/src/main/java/org/eclipse/jkube/maven/plugin/mojo/develop/OpenshiftDeployMojo.java
@@ -44,7 +44,7 @@ public class OpenshiftDeployMojo extends DeployMojo {
 
     @Override
     public File getManifest(KubernetesClient kubernetesClient) {
-        return OpenShift.getOpenShiftManifest(kubernetesClient, super.getManifest(kubernetesClient), openshiftManifest);
+        return OpenShift.getOpenShiftManifest(kubernetesClient, getKubernetesManifest(), openshiftManifest);
     }
 
     @Override