+++ openshift-maven-plugin/plugin/src/main/java/org/eclipse/jkube/maven/plugin/mojo/develop/OpenshiftUndeployMojo.java
@@ -66,7 +66,7 @@ public class OpenshiftUndeployMojo extends UndeployMojo {
 
     @Override
     public File getManifest(KubernetesClient kubernetesClient) {
-        return OpenShift.getOpenShiftManifest(kubernetesClient, super.getManifest(kubernetesClient), openshiftManifest);
+        return OpenShift.getOpenShiftManifest(kubernetesClient, getKubernetesManifest(), openshiftManifest);
     }
 
     @Override