+++ CHANGELOG.md
@@ -24,7 +24,8 @@ Usage:
 * Fix #167: Add CMD for wildfly based applications
 * Added Webapp Wildfly maven quickstart
 * Fix #171: OpenShift pull secret not picked up without registry auth configuration
-* Fix #171: Customized Quarkus application quick start 
+* Fix #171: Customized Quarkus application quick start
+* Fix #101: Removed Openshift specific functionality from Kubernetes Maven Plugin
 
 ### 1.0.0-alpha-2 (2020-04-24)
 * Fix #130: Updated HelmMojo documentation